CHECKPOINT NOTES - Remove sandbox window
There are some weird Gtk issues going on with the note field.

- Sometimes on startup the objects don't have their y-positions
  updated properly, so they are all stacked on top of each other

- The first time you scroll vertically and horizontally the note
  field doesn't redraw itself, but after that it's fine
